Added teachers to groups, and groups to teachers
